Factored public run function from main
The purpose of this change is to expose a public API which has the same
semantic compatability as the CLI itself.

Factored core functionality out of main, and put it in a public function
called `run()`.

This function accepts a struct representing the arguments and returns an
`anyhow::Error`.

Made arguments public & enabled ser/de of it's structs.